Proper packet sending
