+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/PingSpoof.kt
@@ -12,6 +12,7 @@ import net.ccbluex.liquidbounce.event.WorldEvent
 import net.ccbluex.liquidbounce.features.module.Module
 import net.ccbluex.liquidbounce.features.module.ModuleCategory
 import net.ccbluex.liquidbounce.features.module.modules.exploit.PingSpoof.sendPacketsByOrder
+import net.ccbluex.liquidbounce.utils.PacketUtils
 import net.ccbluex.liquidbounce.utils.PacketUtils.handlePacket
 import net.ccbluex.liquidbounce.value.IntegerValue
 import net.minecraft.network.Packet
@@ -56,7 +57,7 @@ object PingSpoof : Module("PingSpoof", ModuleCategory.EXPLOIT) {
         synchronized(packetQueue) {
             packetQueue.entries.removeAll { (packet, timestamp) ->
                 if (all || timestamp <= (System.currentTimeMillis() - delay)) {
-                    handlePacket(packet)
+                    PacketUtils.queuedPackets.add(packet)
                     true
                 } else false
             }
