+++ src/main/java/net/ccbluex/liquidbounce/utils/inventory/InventoryUtils.kt
@@ -135,6 +135,7 @@ object InventoryUtils : MinecraftInstance(), Listenable {
 
     @EventTarget
     fun onPacket(event: PacketEvent) {
+
         if (event.isCancelled) return
 
         when (val packet = event.packet) {
